Add pagination, sorting and styling to airports table.

+++ src/AirportSystem.Domain/Entities/Airport.cs
@@ -10,9 +10,9 @@ namespace AirportSystem.Domain.Entities
 
         public int CityId { get; set; }
 
-        public decimal Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public decimal Longitude { get; set; }
+        public double Longitude { get; set; }
 
         public City City { get; set; }
